extract/requirements.txt
streamlit
python-dotenv
pdfplumber
openai